+++ src/hooks/useAuthorityLinking/useAuthorityLinking.test.js
@@ -82,7 +82,7 @@ describe('Given useAuthorityLinking', () => {
       };
 
       expect(result.current.linkAuthority(authority, authSource, field)).toMatchObject({
-        content: '$a Beethoven, Ludwig van $e composer. $0 http://some.url/n0001 $9 authority-id',
+        content: '$a Beethoven, Ludwig van $e composer. $0 some.url/n0001 $9 authority-id',
         authorityControlledSubfields: ['a', 'b', 't', 'd'],
       });
     });
@@ -103,7 +103,7 @@ describe('Given useAuthorityLinking', () => {
       };
 
       expect(result.current.linkAuthority(authority, authoritySource, field)).toMatchObject({
-        content: '$a authority value $b fakeB $t field for modification $0 http://some.url/n0001 $9 authority-id',
+        content: '$a authority value $b fakeB $t field for modification $0 some.url/n0001 $9 authority-id',
         authorityControlledSubfields: ['a', 'b', 't', 'd'],
       });
     });
@@ -124,7 +124,7 @@ describe('Given useAuthorityLinking', () => {
       };
 
       expect(result.current.linkAuthority(authority, authoritySource, field)).toMatchObject({
-        content: '$a authority value $b fakeB $0 http://some.url/n0001 $9 authority-id $t field for modification',
+        content: '$a authority value $b fakeB $0 some.url/n0001 $9 authority-id $t field for modification',
         authorityControlledSubfields: ['a', 'b', 't', 'd'],
       });
     });
@@ -145,7 +145,7 @@ describe('Given useAuthorityLinking', () => {
       };
 
       expect(result.current.linkAuthority(authority, authoritySource, field)).toMatchObject({
-        content: '$a authority value $b fakeB $e author $e illustrator $t field for modification $0 http://some.url/n0001 $9 authority-id',
+        content: '$a authority value $b fakeB $e author $e illustrator $t field for modification $0 some.url/n0001 $9 authority-id',
         authorityControlledSubfields: ['a', 'b', 't', 'd'],
       });
     });
@@ -183,7 +183,7 @@ describe('Given useAuthorityLinking', () => {
       };
 
       expect(result.current.linkAuthority(authority, authoritySource, field)).toMatchObject({
-        content: '$a authority value $b fakeB $e author $e illustrator $c field for modification $0 http://some.url/n0001 $9 authority-id',
+        content: '$a authority value $b fakeB $e author $e illustrator $c field for modification $0 some.url/n0001 $9 authority-id',
         authorityControlledSubfields: ['a', 'b', 'c'],
       });
     });